+++ test/cronex/job_test.exs
@@ -2,26 +2,7 @@ defmodule Cronex.JobTest do
   use ExUnit.Case
 
   alias Cronex.Job
-
-  defmodule TestDateTime do
-    def start_link do
-      Agent.start_link(fn -> Cronex.DateTime.current end, name: __MODULE__)
-    end
-
-    def set(%Cronex.DateTime{} = date_time) do
-      Agent.update(__MODULE__, fn(_) -> date_time end)
-    end
-
-    def current do
-      Agent.get(__MODULE__, &(&1))
-    end
-  end
-
-  setup_all do
-    TestDateTime.start_link
-    Application.put_env(:cronex, :date_time_provider, __MODULE__.TestDateTime)
-    :ok
-  end
+  alias Cronex.Test
 
   setup do
     {:ok, job_supervisor} = Task.Supervisor.start_link 
@@ -80,10 +61,10 @@ defmodule Cronex.JobTest do
     task = fn -> :ok end
     job = Job.new(:hour, task) 
 
-    TestDateTime.set(%Cronex.DateTime{minute: 0})
+    Test.DateTime.set(%Cronex.DateTime{minute: 0})
     assert true == Cronex.Job.can_run?(job)
 
-    TestDateTime.set(%Cronex.DateTime{minute: 1})
+    Test.DateTime.set(%Cronex.DateTime{minute: 1})
     assert false == Cronex.Job.can_run?(job)
   end
 
@@ -91,16 +72,16 @@ defmodule Cronex.JobTest do
     task = fn -> :ok end
     job = Job.new(:day, task) 
 
-    TestDateTime.set(%Cronex.DateTime{hour: 0, minute: 0})
+    Test.DateTime.set(%Cronex.DateTime{hour: 0, minute: 0})
     assert true == Cronex.Job.can_run?(job)
 
-    TestDateTime.set(%Cronex.DateTime{hour: 0, minute: 1})
+    Test.DateTime.set(%Cronex.DateTime{hour: 0, minute: 1})
     assert false == Cronex.Job.can_run?(job)
 
-    TestDateTime.set(%Cronex.DateTime{hour: 1, minute: 0})
+    Test.DateTime.set(%Cronex.DateTime{hour: 1, minute: 0})
     assert false == Cronex.Job.can_run?(job)
 
-    TestDateTime.set(%Cronex.DateTime{hour: 1, minute: 1})
+    Test.DateTime.set(%Cronex.DateTime{hour: 1, minute: 1})
     assert false == Cronex.Job.can_run?(job)
   end
 
@@ -108,13 +89,13 @@ defmodule Cronex.JobTest do
     task = fn -> :ok end
     job = Job.new(:wednesday, task) 
 
-    TestDateTime.set(%Cronex.DateTime{day_of_week: 3, hour: 0, minute: 0})
+    Test.DateTime.set(%Cronex.DateTime{day_of_week: 3, hour: 0, minute: 0})
     assert true == Cronex.Job.can_run?(job)
 
-    TestDateTime.set(%Cronex.DateTime{day_of_week: 1, hour: 0, minute: 0})
+    Test.DateTime.set(%Cronex.DateTime{day_of_week: 1, hour: 0, minute: 0})
     assert false == Cronex.Job.can_run?(job)
 
-    TestDateTime.set(%Cronex.DateTime{day_of_week: 3, hour: 1, minute: 0})
+    Test.DateTime.set(%Cronex.DateTime{day_of_week: 3, hour: 1, minute: 0})
     assert false == Cronex.Job.can_run?(job)
   end
 
@@ -122,13 +103,13 @@ defmodule Cronex.JobTest do
     task = fn -> :ok end
     job = Job.new(:month, task) 
 
-    TestDateTime.set(%Cronex.DateTime{day: 1, hour: 0, minute: 0})
+    Test.DateTime.set(%Cronex.DateTime{day: 1, hour: 0, minute: 0})
     assert true == Cronex.Job.can_run?(job)
 
-    TestDateTime.set(%Cronex.DateTime{day: 2, hour: 0, minute: 0})
+    Test.DateTime.set(%Cronex.DateTime{day: 2, hour: 0, minute: 0})
     assert false == Cronex.Job.can_run?(job)
 
-    TestDateTime.set(%Cronex.DateTime{day: 1, hour: 1, minute: 0})
+    Test.DateTime.set(%Cronex.DateTime{day: 1, hour: 1, minute: 0})
     assert false == Cronex.Job.can_run?(job)
   end
 
@@ -136,13 +117,13 @@ defmodule Cronex.JobTest do
     task = fn -> :ok end
     job = Job.new(:year, task) 
 
-    TestDateTime.set(%Cronex.DateTime{month: 1, day: 1, hour: 0, minute: 0})
+    Test.DateTime.set(%Cronex.DateTime{month: 1, day: 1, hour: 0, minute: 0})
     assert true == Cronex.Job.can_run?(job)
 
-    TestDateTime.set(%Cronex.DateTime{month: 2, day: 1, hour: 0, minute: 0})
+    Test.DateTime.set(%Cronex.DateTime{month: 2, day: 1, hour: 0, minute: 0})
     assert false == Cronex.Job.can_run?(job)
 
-    TestDateTime.set(%Cronex.DateTime{month: 1, day: 2, hour: 0, minute: 0})
+    Test.DateTime.set(%Cronex.DateTime{month: 1, day: 2, hour: 0, minute: 0})
     assert false == Cronex.Job.can_run?(job)
   end
 end